Changes to light, normal mapping, and the gl canvas
Three things:
1. The gl canvas was transluscent, which caused problematic "ghosting"
from 2d into gl.
2. The normal map was all gunked up. Properly testing it w/r/t the vertex
shader helped get a clearer sense of the thing. Worth noting also that a
lot of the normal map depends on lhs/rhs differentiation. Probably will
also need to account for top and bottom halves but the light source in
this project should probably always come from above.
3. Lighting was all gunked up too. It was too close, too centered,
the diffuse mult was off, and so were the specular and shininess data
points.